move modules to lib
