+++ src/lib/corpora/characters.json
