+++ src/lib/corpora/emotions.json
