+++ src/lib/corpora/fabrics.json
