+++ src/lib/corpora/industries.json
