+++ src/lib/tarot/Tarot_Card_Detail.svelte
