+++ src/lib/tarot/Tarot_Card_Thumbnail.svelte
