+++ src/lib/tarot/Tarot_Thumbnail.svelte
