+++ src/lib/zodiac/Aquarius.svelte
