+++ src/lib/zodiac/Capricorn.svelte
