+++ src/lib/zodiac/Taurus.svelte
