+++ src/lib/zodiac/Virgo.svelte
