+++ src/routes/tarot/index.svelte
@@ -3,12 +3,12 @@
 	import Footer from '$lib/Footer.svelte';
 	import {last} from '@feltcoop/felt/util/array.js';
 
-	import Tarot_Card_Button from '../../tarot/Tarot_Card_Button.svelte';
-	import Tarot_Card_Detail from '../../tarot/Tarot_Card_Detail.svelte';
-	import Drawn_Tarot_Cards from '../../tarot/Drawn_Tarot_Cards.svelte';
-	import {cards as cards_data} from '../../tarot/tarot.json';
-	import {draw_cards, TAROT_CARD_MIN_WIDTH, TAROT_CARD_MIN_HEIGHT} from '../../tarot/tarot';
-	import type {Tarot_Card} from '../../tarot/tarot.js';
+	import Tarot_Card_Button from '$lib/tarot/Tarot_Card_Button.svelte';
+	import Tarot_Card_Detail from '$lib/tarot/Tarot_Card_Detail.svelte';
+	import Drawn_Tarot_Cards from '$lib/tarot/Drawn_Tarot_Cards.svelte';
+	import {cards as cards_data} from '$lib/tarot/tarot.json';
+	import {draw_cards, TAROT_CARD_MIN_WIDTH, TAROT_CARD_MIN_HEIGHT} from '$lib/tarot/tarot';
+	import type {Tarot_Card} from '$lib/tarot/tarot.js';
 	import {shuffle} from '$lib/random';
 
 	// TODO refactor, extract some components
